src/requirements.txt
torch-directml
numpy
Pillow
opencv-python
matplotlib
tqdm
